chore(trace): move processors
