+++ ddtrace/appsec/_iast/processor.py
@@ -2,13 +2,13 @@ from typing import TYPE_CHECKING
 
 import attr
 
+from ddtrace._trace.processor.trace import SpanProcessor
 from ddtrace.appsec._constants import APPSEC
 from ddtrace.appsec._constants import IAST
 from ddtrace.constants import ORIGIN_KEY
 from ddtrace.ext import SpanTypes
 from ddtrace.internal import core
 from ddtrace.internal.logger import get_logger
-from ddtrace.internal.processor import SpanProcessor
 
 from .._trace_utils import _asm_manual_keep
 from . import oce
