+++ ddtrace/appsec/_processor.py
@@ -13,6 +13,7 @@ from typing import Set
 from typing import Tuple
 from typing import Union
 
+from ddtrace._trace.processor.trace import SpanProcessor
 from ddtrace.appsec import _asm_request_context
 from ddtrace.appsec._capabilities import _appsec_rc_file_is_not_static
 from ddtrace.appsec._constants import APPSEC
@@ -32,7 +33,6 @@ from ddtrace.constants import RUNTIME_FAMILY
 from ddtrace.ext import SpanTypes
 from ddtrace.internal import core
 from ddtrace.internal.logger import get_logger
-from ddtrace.internal.processor import SpanProcessor
 from ddtrace.internal.rate_limiter import RateLimiter
 from ddtrace.settings.asm import config as asm_config
 from ddtrace.span import Span
