+++ ddtrace/filters.py
@@ -5,8 +5,8 @@ from typing import List  # noqa:F401
 from typing import Optional  # noqa:F401
 from typing import Union  # noqa:F401
 
+from ddtrace._trace.processor.trace import TraceProcessor
 from ddtrace.ext import http
-from ddtrace.internal.processor.trace import TraceProcessor
 
 
 if TYPE_CHECKING:  # pragma: no cover
