+++ ddtrace/internal/processor/endpoint_call_counter.py
@@ -2,10 +2,10 @@ import typing
 
 import attr
 
+from ddtrace._trace.processor.trace import SpanProcessor
 from ddtrace.ext import SpanTypes
 from ddtrace.internal import forksafe
 from ddtrace.internal.compat import ensure_text
-from ddtrace.internal.processor import SpanProcessor
 from ddtrace.span import Span  # noqa:F401
