+++ ddtrace/internal/processor/stats.py
@@ -8,6 +8,7 @@ from ddsketch.pb.proto import DDSketchProto
 
 import ddtrace
 from ddtrace import config
+from ddtrace._trace.processor.trace import SpanProcessor
 from ddtrace.internal import compat
 from ddtrace.internal.utils.retry import fibonacci_backoff_with_jitter
 from ddtrace.span import _is_top_level
@@ -21,7 +22,6 @@ from ..hostname import get_hostname
 from ..logger import get_logger
 from ..periodic import PeriodicService
 from ..writer import _human_size
-from . import SpanProcessor
 
 
 if typing.TYPE_CHECKING:  # pragma: no cover
